+++ R/metacheck-package.R
@@ -2,7 +2,6 @@
 "_PACKAGE"
 
 ## usethis namespace: start
-#' @import svutils
 ## Suppress R CMD check note for packages used only in modules
 #' @importFrom DT datatable
 ## usethis namespace: end