+++ tests/testthat/test-module-.R
@@ -24,7 +24,9 @@ test_that("module_find", {
   expect_error(module_find("notamodule"),
                "There were no modules that matched notamodule",
                fixed = TRUE )
-
+  log <- lastlog()
+  expect_equal(log$module, "notamodule")
+  expect_equal(log$label, "module_find")
 
   # find built-in modules
   builtin <- module_list()