+++ NEWS.md
@@ -1,3 +1,11 @@
+# metacheck 0.0.0.9106
+
+* Added functions from svutils back in. 
+* Reorganised some ML read-in functions (internal).
+* Ollama further support in `llm()` and vignette.
+* The code_check module handles local file with the argument `local_path`
+* New `local_files()` function
+
 # metacheck 0.0.0.9105
 
 * Much less buggy `.grobid_to_bibr()` conversion, handling URLs in text, xrefs, url, and eq tables better.